move client features to separate file
